% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbpalette.R
\docType{data}
\name{cbpalette}
\alias{cbpalette}
\title{Default color palette}
\format{
An object of class \code{character} of length 15.
}
\usage{
cbpalette
}
\description{
Default color palette
}
\keyword{datasets}
