% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_calc_growth_trans.R
\name{gmacs_calc_growth_trans}
\alias{gmacs_calc_growth_trans}
\title{Calculate Growth Transition Matrix from Parameters}
\usage{
gmacs_calc_growth_trans(gpars, n_size_bins, size_breaks, type)
}
\arguments{
\item{gpars}{Growth parameters list in same order as .ctl file.}

\item{n_size_bins}{Number of size classes.}

\item{size_breaks}{Breaks between size classes, vector of length \code{n_size_bins + 1}.}

\item{type}{Growth parameterization corresponding to GMACS option. Currently only type = 7 is available.}
}
\value{
Growth transition matrix
}
\description{
Produces growth transition matrix provided parameters and dimensions
}
\examples{
gmacs_calc_growth_trans(gpars = list(161, 0.35, 0.05, 0.05), n_size_bins = 13, size_breaks = seq(40, 170, 10), type = 7)

}
