% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_carry_forward.R
\name{gmacs_carry_forward}
\alias{gmacs_carry_forward}
\title{Carry forward}
\usage{
gmacs_carry_forward(x, group_col = "year")
}
\arguments{
\item{x}{Tibble.}

\item{group_col}{Name of column to extend forward, usually year.}
}
\value{
Tibble is same format as x, with one more year or whatever group
}
\description{
Extend values from terminal year to next (e.g., survey selectivity in last year)
}
\examples{
macs_get_slx(list(all_out)) \%>\% gmacs_carry_forward()

}
