% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_do_jitter.R
\name{gmacs_do_jitter}
\alias{gmacs_do_jitter}
\title{Run Jitter Analysis}
\usage{
gmacs_do_jitter(
  gmacs.dat,
  jitter_type = 1,
  jitter_use_pin = 0,
  sd,
  iter,
  save_csv = T,
  csv_dir = NULL,
  save_plot = T,
  plot_dir = NULL,
  plot_only = F,
  batch = F,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{gmacs.dat}{File path to gmacs.dat file.}

\item{jitter_type}{Method used for jittering: 1) asymmetric normal within bounds, 2) uniform with shrunken interval around initial value, 3) normal with logistic transformation}

\item{jitter_use_pin}{Use .PIN file for jittered starting values.}

\item{sd}{Jitter standard deviation.}

\item{iter}{Number of jittering runs.}

\item{save_csv}{T/F save csv file output. Default = T.}

\item{csv_dir}{Null. Directory in which to save output. If NULL, a directory called 'output' will be created in the same directory as gmacs.dat.}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{plot_only}{T/F only plot results (analysis already run). Default = F.}

\item{batch}{T/F Run jitter iterations in batch file, skip reading and saving results. Default = F.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.16".}
}
\value{
A subdirectory called 'jitter' will be created in the same directory as the gmacs.dat file. Within './jitter' a subdirectory corresponding to each run will be created. The output and plot will be saved in the provided directories or output as a list if save_csv and save_plot = F.
}
\description{
Run GMACS jitter analysis, then save and plot results.
}
\examples{
gmacs_do_jitter(gmacs.dat = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat", sd = 0.3, iter = 100)

}
