% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_do_ofl_dist.R
\name{gmacs_do_ofl_dist}
\alias{gmacs_do_ofl_dist}
\title{Estimate Distribution of OFL}
\usage{
gmacs_do_ofl_dist(
  gmacs.dat,
  n_replicates,
  n_draws,
  wait = T,
  save_plot = T,
  plot_dir = NULL,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{gmacs.dat}{File path to gmacs.dat file.}

\item{n_replicates}{Length of MCMC chain.}

\item{n_draws}{Number of draws to save.}

\item{wait}{Passed to shell(): a logical (not NA) indicating whether the R interpreter should wait for the command to finish, or run it asynchronously. Default = T.}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.16".}
}
\value{
A subdirectory called 'ofl_distribution' will be created in the same directory as the gmacs.dat file. Within './ofl_distribution' the model version with MCMC will be run. The plot will be saved in the provided directory or output as a list include mcoutREF if save_plot = F.
}
\description{
Run GMACS with MCMC of reference point calculation.
}
\examples{
gmacs_do_ofl_dist(gmacs.dat = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat",n_replicates = 1000000, n_draws = 1000)

}
