% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_do_retrospective.R
\name{gmacs_do_retrospective}
\alias{gmacs_do_retrospective}
\title{Run Retrospective Analysis}
\usage{
gmacs_do_retrospective(
  gmacs.dat,
  n_peel,
  pin = F,
  wait = T,
  plot_mmb = T,
  save_plot = T,
  plot_dir = NULL,
  plot_only = F,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{gmacs.dat}{File path to gmacs.dat file.}

\item{n_peel}{Number of retrospective peels.}

\item{pin}{T/F use pin file. Default = F.}

\item{wait}{Passed to shell(): a logical (not NA) indicating whether the R interpreter should wait for the command to finish, or run it asynchronously. Default = T.}

\item{plot_mmb}{T/F make plot of MMB by peel.}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{plot_only}{T/F only plot results (analysis already run). Default = F.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.16".}
}
\value{
A subdirectory called 'retrospectives' will be created in the same directory as the gmacs.dat file. Within './retrospectives' a subdirectory corresponding to each peel will be created. The plot will be saved in the provided directory or output as a list if save_plot = F.
}
\description{
Run GMACS retrospective analysis, then save and plot results.
}
\examples{
gmacs_do_retrospective(gmacs.dat = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat", n_peel = 10)

}
