% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_get_esb.R
\name{gmacs_get_esb}
\alias{gmacs_get_esb}
\title{Get Effective Spawning Biomass}
\usage{
gmacs_get_esb(all_out = NULL, file = NULL, model_name = NULL, version = NULL)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}
}
\value{
Tibble with year and effective spawning biomass.
}
\description{
Compute effective spawning biomass (Zheng 1997) used in BBRKC harvest strategy
}
\examples{
gmacs_get_esb(all_out = list(mod_23.0a, mod_23.1b))

}
