% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_get_recruitment_distribution.R
\name{gmacs_get_recruitment_distribution}
\alias{gmacs_get_recruitment_distribution}
\title{Get Recruitment Distribution}
\usage{
gmacs_get_recruitment_distribution(
  all_out = NULL,
  n_rec_class = NULL,
  file = NULL,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{n_rec_class}{Number of recruitment classes as list with first element vector for males, second element vector for females.}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}
}
\value{
Tibble of proportion by size class and sex with column 'model'.
}
\description{
Compute size distribution of recruit classes
}
\examples{
gmacs_get_lik_type_pen(all_out = list(mod_23.0a, mod_23.1b))

}
