% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_get_tac.R
\name{gmacs_get_tac}
\alias{gmacs_get_tac}
\title{Get Catch Summary}
\usage{
gmacs_get_tac(
  all_out = NULL,
  file = NULL,
  model_name = NULL,
  version = NULL,
  stock = c("AIGKC", "BBRKC", "BSTC"),
  hs_pars = NULL,
  bbrkc_lba = F
)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}

\item{stock}{name of stock to apply harvest strategy, currently only AIGKC}

\item{hs_pars}{list of harvest strategy parameters needed. For AIGKC example: list(max_ex = 0.15, avg_wt = 4.41, calc_wt = F)}
}
\value{
Tibble of tac summary from.
}
\description{
Isolate catch summary data from Gmacsall.out by model
}
\examples{
gmacs_get_catch_summary(all_out = list(mod_23.0a, mod_23.1b))

}
