% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_par2pin.R
\name{gmacs_par2pin}
\alias{gmacs_par2pin}
\title{Write .pin file}
\usage{
gmacs_par2pin(par)
}
\arguments{
\item{par}{path to .par file.}
}
\value{
File saved in specified location.
}
\description{
Create GMACS .pin file in R from existing par file
}
\examples{
gmacs_par2pin(input, file = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.par")

}
