% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_plot_data_range.R
\name{gmacs_plot_data_range}
\alias{gmacs_plot_data_range}
\title{Plot Data Range}
\usage{
gmacs_plot_data_range(
  all_out = NULL,
  save_plot = T,
  plot_dir = NULL,
  file = NULL,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}
}
\value{
Plot of data range by type
}
\description{
Plot data rangey for a single model
}
\examples{
gmacs_plot_data_range(all_out = list(model_25.0))

}
