% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_plot_f.R
\name{gmacs_plot_f}
\alias{gmacs_plot_f}
\title{Plot Fishing Mortality}
\usage{
gmacs_plot_f(
  all_out = NULL,
  save_plot = T,
  plot_dir = NULL,
  yrs = NULL,
  data_summary = NULL,
  file = NULL,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{yrs}{NULL. Subset a specific year range, example: c(1990:2022)}

\item{data_summary}{NULL. Alternate way to bring in data, output of gmacs_get_catch_summary()}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}
}
\value{
Plot of fishing mortality by sex and fleet
}
\description{
Plot F by sex and fleet from Gmacsall.out data summary
}
\examples{
gmacs_plot_f(all_out = list(model_25.0))

}
