% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_plot_mean_size.R
\name{gmacs_plot_mean_size}
\alias{gmacs_plot_mean_size}
\title{Plot Fit to Mean Size Comp Data}
\usage{
gmacs_plot_mean_size(
  all_out = NULL,
  save_plot = T,
  plot_dir = NULL,
  size_lab = "Size",
  agg_series = T,
  agg_series_label = NULL,
  data_summary = NULL,
  file = NULL,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{size_lab}{Optional, custom size axis label, as character vector, example: "Carapace Length (mm)". Default = "Size".}

\item{agg_series}{T/F Plot aggregate series together. Default = T.}

\item{agg_series_label}{character vector of labels for aggregate series, ex: c("Male", "Female") or c("New Shell", "Old Shell") or list with elements being character vectors for each aggregate series (if you want to use different labels)}

\item{data_summary}{NULL. Alternate way to bring in data, output of gmacs_get_catch_summary()}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}
}
\value{
Plot of fit to mean size data by series
}
\description{
Plot mean size based on fits to size data from Gmacsall.out data summary
}
\examples{
gmacs_plot_mean_size(list(bbrkc), save_plot = T, size_lab = "CL",  agg_series = F)

}
