% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_plot_recruitment_distribution.R
\name{gmacs_plot_recruitment_distribution}
\alias{gmacs_plot_recruitment_distribution}
\title{Plot Recruitment Distribution}
\usage{
gmacs_plot_recruitment_distribution(
  all_out = NULL,
  save_plot = T,
  plot_dir = NULL,
  n_rec_class = NULL,
  size_lab = "Size",
  data_summary = NULL,
  file = NULL,
  model_name = NULL,
  version = NULL
)
}
\arguments{
\item{all_out}{Output from gmacs_read_allout() as nested list. Example: all_out = list(mod_23.0a, mod_23.1b).}

\item{save_plot}{T/F save plot. Default = T.}

\item{plot_dir}{Null. Directory in which to save plot. If NULL, a directory called 'plots' will be created in the same directory as gmacs.dat.}

\item{n_rec_class}{Number of recruitment classes as list with first element vector for males, second element vector for females.}

\item{size_lab}{Optional, custom size axis label, as character vector, example: "Carapace Length (mm)". Default = "Size".}

\item{data_summary}{NULL. Alternate way to bring in data, output of gmacs_get_catch_summary()}

\item{file}{NULL. File paths to Gmacsall.out for each model to compare, passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{model_name}{NULL. Character string passed to gmacs_read_allout(). Expressed as character vector, not needed if all.out is provided.}

\item{version}{NULL. Character string passed to gmacs_read_allout() denoting GMACS version, not needed if all.out is provided.}
}
\value{
Plot of recruitment distribution.
}
\description{
Plot recruitment distribution from gmacs_get_recruitment_distribution()
}
\examples{
gmacs_plot_recruitment_distribution(all_out = list(model_25.0), n_rec_class = 5)

}
