% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_read_allout.R
\name{gmacs_read_allout}
\alias{gmacs_read_allout}
\title{Read Gmacsall.out file}
\usage{
gmacs_read_allout(file, model_name = NULL, version = NULL)
}
\arguments{
\item{file}{NULL. Path to Gmacsall.out file.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.16".}
}
\value{
List of file contents and model results
}
\description{
Load Gmacsall.out file in R
}
\examples{
gmacs_read_allout(file = "./AIGKC/models/2024/may/EAG/23.1b/Gmacsall.out", model_name = "23.1b")

}
