% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_read_ctl.R
\name{gmacs_read_ctl}
\alias{gmacs_read_ctl}
\title{Read .ctl file}
\usage{
gmacs_read_ctl(ctl_file, dat_file, version = NULL)
}
\arguments{
\item{ctl_file}{NULL. Path to .ctl file.}

\item{dat_file}{NULL. Path to .dat file.}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.17".}
}
\value{
List of .ctl file contents. LIKELY STILL CONTAINS BUGS
}
\description{
Load .ctl file in R
}
\examples{
gmacs_read_ctl(ctl_file = "./AIGKC/models/2024/may/EAG/23.1b/EAG_23_1b.ctl", dat_file = "./AIGKC/models/2024/may/EAG/23.1b/EAG_23_1b.dat", model_name = "23.1b")

}
