% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_read_files_dat.R
\name{gmacs_read_files_dat}
\alias{gmacs_read_files_dat}
\title{Read gmacs.dat file}
\usage{
gmacs_read_files_dat(gmacs.dat, model_name = NULL, version = NULL)
}
\arguments{
\item{gmacs.dat}{NULL. Path to gmacs.dat file.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.17".}
}
\value{
List of gmacs.dat file contents.
}
\description{
Load GMACS gmacs.dat file in R
}
\examples{
gmacs_read_files_dat(dat_file = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat", model_name = "23.1b")

}
