% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_read_mcoutREF.R
\name{gmacs_read_mcoutREF}
\alias{gmacs_read_mcoutREF}
\title{Read mcoutREF file}
\usage{
gmacs_read_mcoutREF(file, model_name = NULL, version = NULL)
}
\arguments{
\item{file}{NULL. Path to mcoutREF.rep file.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.16".}
}
\value{
Tibble of mcoutREF.rep file contents.
}
\description{
Load projection reference point results file in R
}
\examples{
gmacs_read_mcoutREF(file = "./AIGKC/models/2024/may/EAG/23.1b/mcoutREF.rep", model_name = "23.1b")

}
