% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_read_std.R
\name{gmacs_read_std}
\alias{gmacs_read_std}
\title{Read gmacs.std file}
\usage{
gmacs_read_std(file, model_name = NULL, sub_text = NULL)
}
\arguments{
\item{file}{NULL. Path to gmacs.std file.}

\item{model_name}{NULL. Character string to save as object in output, later to be used for plot legends. Example: "23.1b".}

\item{sub_text}{NULL. Character string filtering, example: "ssb".}
}
\value{
Tibble containing .std file contents
}
\description{
Load gmacs.std file in R
}
\examples{
gmacs_read_std(gmacs.dat = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.std", model_name = "23.1b", sub_text = "ssb")

}
