% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_write_dat.R
\name{gmacs_write_dat}
\alias{gmacs_write_dat}
\title{Write .dat file}
\usage{
gmacs_write_dat(input, file = NULL, version = NULL)
}
\arguments{
\item{input}{Named list of data needed to write file (need more info here), look at output of gmacs_read_dat()}

\item{file}{NULL. File path to write .dat file}

\item{version}{NULL. Character string denoting GMACS version. Default: "2.20.31".}
}
\value{
File saved in specified location.
}
\description{
Write GMACS .dat file in R
}
\examples{
gmacs_write_dat(input, file = "./AIGKC/models/2024/may/EAG/23.1b/EAG_23_1b.dat")

}
