% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmacs_write_files_dat.R
\name{gmacs_write_files_dat}
\alias{gmacs_write_files_dat}
\title{Write gmacs.dat file}
\usage{
gmacs_write_files_dat(input, file)
}
\arguments{
\item{input}{Named list of data needed to write file (need more info here), look at output of gmacs_read_files_dat()}

\item{file}{NULL. File path to write gmacs.dat file}
}
\value{
File saved in specified location.
}
\description{
Write GMACS gmacs.dat file in R
}
\examples{
gmacs_write_dat(input, file = "./AIGKC/models/2024/may/EAG/23.1b/gmacs.dat")

}
