% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickr.R
\name{tickr}
\alias{tickr}
\title{Get minor ticks}
\usage{
tickr(data, var, to)
}
\arguments{
\item{data}{Data frame with column for year}

\item{var}{Name of year column}

\item{to}{Interval}
}
\value{
Tibble with breaks and labels.
}
\description{
Get minor ticks breaks and labels for axis scale_ functions
}
\examples{
tickr(tibble(yr = 1900:2100), yr, 5)

}
