% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yraxis.R
\docType{data}
\name{yraxis}
\alias{yraxis}
\title{Default ear axis minor ticks in 5-yr interval}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 201 rows and 2 columns.
}
\usage{
yraxis
}
\description{
Default ear axis minor ticks in 5-yr interval
}
\keyword{datasets}
